Fix data types of the currency fields.

+++ synlogistics/main/models.py
@@ -69,7 +69,7 @@ class Account(models.Model):
 	description = models.CharField(max_length=765, blank=True)
 	account_type = models.IntegerField(null=False, blank=False, choices=TYPE_CHOICES)
 	is_readonly = models.BooleanField(null=False, default=False)
-	_balance = models.FloatField(null=False, blank=False, default=0.0)
+	_balance = models.DecimalField(decimal_places=5, null=False, blank=False, default=0.0)
 	parent = models.ForeignKey('self', null=True, related_name='children')
 
 	def get_balance(self):
@@ -105,7 +105,7 @@ class Transaction(models.Model):
 	related = models.ManyToManyField('self', blank=False)
 	relation = models.ForeignKey(Relation, null=True)
 	description = models.CharField(max_length=765, blank=True)
-	amount = models.FloatField(null=True, blank=True)
+	amount = models.DecimalField(decimal_places=5, null=True, blank=True)
 	invoice_number = models.CharField(max_length=45, null=True, blank=True)
 	purchase_order = models.ForeignKey(PurchaseOrder, null=True)
 	#sale = models.ForeignKey('SaleItem')
@@ -127,7 +127,7 @@ class Location(models.Model):
 
 class Vat(models.Model):
 	name = models.CharField(max_length=30, blank=True)
-	percent = models.FloatField(null=True, blank=True)
+	percent = models.DecimalField(decimal_places=5, null=True, blank=True)
 	class Meta:
 		db_table = u'vat'
 
@@ -170,8 +170,8 @@ class InternalOrder(models.Model):
 	product = models.ForeignKey(Product)
 	product_group = models.ForeignKey(ProductGroup)
 	product_description = models.CharField(max_length=300, blank=True)
-	purchase_price_indication = models.FloatField(null=True, blank=True)
-	selling_price = models.FloatField(null=True, blank=True)
+	purchase_price_indication = models.DecimalField(decimal_places=5, null=True, blank=True)
+	selling_price = models.DecimalField(decimal_places=5, null=True, blank=True)
 	supplier_suggestion = models.ForeignKey(Relation, related_name='+')
 	for_customer = models.BooleanField(default=False)
 	customer = models.ForeignKey(Relation)
@@ -183,7 +183,7 @@ class PurchaseOrderItem(models.Model):
 	purchase_order = models.ForeignKey(PurchaseOrder)
 	internal_order = models.ForeignKey(InternalOrder)
 	state = models.IntegerField(null=True, blank=True)
-	purchase_price = models.FloatField(null=True, blank=True)
+	purchase_price = models.DecimalField(decimal_places=5, null=True, blank=True)
 	order_amount = models.IntegerField(null=True, blank=True)
 	units_per_pack = models.IntegerField(null=True, blank=True)
 	subproduct = models.ForeignKey(Subproduct)
@@ -245,7 +245,7 @@ class InventoryItem(models.Model):
 	inventory = models.ForeignKey('Inventory')
 	item = models.ForeignKey(Item)
 	location = models.ForeignKey(Location)
-	value = models.FloatField(null=True, blank=True)
+	value = models.DecimalField(decimal_places=5, null=True, blank=True)
 	class Meta:
 		db_table = u'inventory_data'
 
@@ -254,7 +254,7 @@ class ProductSellingprice(models.Model):
 	product = models.ForeignKey(Product)
 	commencing_date = models.DateField(null=True, blank=True)
 	set_date = models.DateField(null=True, blank=True)
-	price = models.FloatField(null=True, blank=True)
+	price = models.DecimalField(decimal_places=5, null=True, blank=True)
 	class Meta:
 		db_table = u'product_sellingprices'
